# Conflicts:
#	front/src/BoardEditor/BoardEditor.js
#	front/src/Community/Community.js
#	front/src/Room/Room.js

+++ back/tcgm_web_app/lib/tcgm_web_app/board_zones/board_zone.ex
@@ -6,11 +6,11 @@ defmodule TcgmWebApp.BoardZones.BoardZone do
 
   schema "board_zones" do
     field :name, :string
-    field :width, :integer
-    field :height, :integer
-    field :x, :integer
-    field :y, :integer
-    field :border_radius, :integer
+    field :width, :float
+    field :height, :float
+    field :x, :float
+    field :y, :float
+    field :border_radius, :float
     field :background_image, :string
 
     belongs_to :board, TcgmWebApp.Boards.Board
